requirements.txt
youtube-transcript-api==0.6.3
langchain-core==0.3.35
langchain-community==0.3.17
langchain-openai==0.3.6
langchain-text-splitters==0.3.6
streamlit==1.41.1
watchdog==6.0.0
pytest==8.3.4
peewee==3.17.9
python-dotenv==1.0.1
chromadb==0.6.3
langchain-chroma==0.2.2
randomname==0.2.1
tiktoken==0.8.0
openai-whisper==20240930
pytubefix==8.12.1
transformers==4.35.0
